card

+++ server/routes/card.js
@@ -4,7 +4,7 @@ const VALID = require('../constant/valid');
 const router = require('express').Router();
 const { body, validationResult } = require('express-validator');
 
-const { Card } = require('../models');
+const { sequelize, Card } = require('../models');
 
 const valid = [
     body('title')
@@ -14,11 +14,18 @@ const valid = [
 
 
 
-router.post('/list', async(req, res, next) => {
+router.post(['/', '/list'], async(req, res, next) => {
 
     try {
 
-        await Card.findAll().then( list => {
+        const cno = req.params.cno;
+
+        await Card.findAll({
+            where: {
+                uno: req.user.uno,
+            }
+
+        }).then( list => {
 
             res.json({
                 api: req.originalUrl,
@@ -54,7 +61,10 @@ router.post('/write', valid, async(req, res, next) => {
         const card = {...req.body};
         card.uno = req.user.uno;
 
-        await Card.create(card).then( card => {
+        await Card.create(
+            card
+
+        ).then( card => {
 
             res.json({
                 api: req.originalUrl,
@@ -63,7 +73,6 @@ router.post('/write', valid, async(req, res, next) => {
                     code: RESULT.SUCCESS.code,
                     message: RESULT.SUCCESS.message,
                 },
-                card: card.dataValues,
             });
 
         }).catch( err => {
@@ -76,9 +85,14 @@ router.post('/write', valid, async(req, res, next) => {
 
 });
 
-router.post('/modify', valid, async(req, res, next) => {
+router.post(['/modify', '/modify:cno'], valid, async(req, res, next) => {
 
     try {
+
+        const cno = req.body.cno || req.params.cno;
+
+        if(!cno)
+            return next(new Error('게시물 번호를 입력해주세요.'));
 
         const vaildError = validationResult(req);
 
@@ -87,24 +101,37 @@ router.post('/modify', valid, async(req, res, next) => {
             return next(new Error('유효성 오류'));
         }
 
-        const card = {...req.body};
+        await Card.findOne({
+            where: sequelize.and({
+                cno,
+                uno: req.user.uno,
+            }),
 
-        await Card.update( card, {
-            where : { cno: card.cno }
+        }).then( card => {
 
-        }).then( result => {
+            if(!card)
+                return next('해당하는 게시물이 없습니다.');
 
-            res.json({
-                api: req.originalUrl,
-                result: {
-                    boolean: RESULT.SUCCESS.boolean,
-                    code: RESULT.SUCCESS.code,
-                    message: RESULT.SUCCESS.message,
-                },
+            Card.update( {
+                ...req.body
+            }, {
+                where : { cno }
+
+            }).then( result => {
+
+                res.json({
+                    api: req.originalUrl,
+                    result: {
+                        boolean: RESULT.SUCCESS.boolean,
+                        code: RESULT.SUCCESS.code,
+                        message: RESULT.SUCCESS.message,
+                    },
+                });
+
+            }).catch( err => {
+                next(err);
             });
 
-        }).catch( err => {
-            next(err);
         });
 
     } catch(err){
@@ -113,25 +140,52 @@ router.post('/modify', valid, async(req, res, next) => {
 
 });
 
-router.post('/delete', async(req, res, next) => {
+router.post(['/delete', '/delete/:cno'], async(req, res, next) => {
 
     try {
 
-        await Card.destroy({
-            where : { cno: req.body.cno }
+        const cno = req.body.cno || req.params.cno;
 
-        }).then( result => {
+        if(!cno)
+            return next(new Error('게시물 번호를 입력해주세요.'));
 
-            res.json({
-                api: req.originalUrl,
-                result: {
-                    boolean: RESULT.SUCCESS.boolean,
-                    code: RESULT.SUCCESS.code,
-                    message: RESULT.SUCCESS.message,
-                },
+        const t = await sequelize.transaction();
+
+        await Card.findAll({
+            where: sequelize.and({
+                cno,
+                uno: req.user.uno,
+            }),
+        },{
+            transaction: t,
+
+        }).then( card => {
+
+            if(!card || Object.keys(card).length == 0)
+                return next('해당하는 게시물이 없습니다.');
+
+            return Card.destroy({
+                where : { cno }
+            },{
+                transaction: t,
+
+            }).then( result => {
+
+                res.json({
+                    api: req.originalUrl,
+                    result: {
+                        boolean: RESULT.SUCCESS.boolean,
+                        code: RESULT.SUCCESS.code,
+                        message: RESULT.SUCCESS.message,
+                    },
+                });
+
+                t.commit();
+
             });
 
         }).catch( err => {
+            t.rollback();
             next(err);
         });
